refactor(api): migrate Azure Functions to TypeScript
